Add Alghemy#overwrite functionality to Cdp

+++ lib/alghemy/rubrics/cdp.rb
@@ -1,16 +1,20 @@
 require 'alghemy/ancestors'
+require 'alghemy/methods'
 
 module Alghemy
   module Rubrics
     class Cdp < Ancestors[:rubric]
+      include Methods[:alget]
+
       def option_templates
         stuff[:option_templates]
       end
 
       def invoke( io )
         previous_file = io[:output].to_s
-        super unless File.exists?(previous_file)
-        File.delete previous_file
+        if File.exists?(previous_file) && alget(:overwrite)
+          File.delete previous_file
+        end
         super
       end
     end
